Revert "Don't remove infobar on editing - fixing the cell does it for us."
This isn't actually as great as I thought it would be - it's kinda ugly for the infobar
to remain while we're editing. The issue remains that if the user cancels editing, the infobar
is gone but the cell hasn't been fixed. Otoh the user is actively shooting themselves in the foot
by doing that so do we really need to fix it?

This reverts commit d1af32aeb95530979e3944cc6c43060264d6db88.

Change-Id: Ib952fb751316e693a5d9144eddc8cd636960370f